refactor(debug): use dynamic import for subcommand
