--- packages/opencode/src/cli/cmd/debug/scrap.ts
@@ -1,16 +1,0 @@
-import { EOL } from "os"
-import { Project } from "@/project/project"
-import * as Log from "@opencode-ai/core/util/log"
-import { cmd } from "../cmd"
-
-export const ScrapCommand = cmd({
-  command: "scrap",
-  describe: "list all known projects",
-  builder: (yargs) => yargs,
-  async handler() {
-    const timer = Log.Default.time("scrap")
-    const list = await Project.list()
-    process.stdout.write(JSON.stringify(list, null, 2) + EOL)
-    timer.stop()
-  },
-})
