--- packages/opencode/src/cli/cmd/debug/startup.ts
@@ -1,11 +1,0 @@
-import { EOL } from "os"
-import { cmd } from "../cmd"
-
-export const StartupCommand = cmd({
-  command: "startup",
-  describe: "print startup timing",
-  builder: (yargs) => yargs,
-  handler() {
-    process.stdout.write(performance.now().toString() + EOL)
-  },
-})
